Added Stats Route, Rewrote delete
* Rewrote delete function to allow for better usage with new looking
glass function
* Moved some logic to mod file, will add more to it later.

+++ src/main.rs
@@ -1,8 +1,11 @@
 use actix_files::NamedFile;
-use actix_web::{web, FromRequest, HttpResponse};
+use actix_web::{web, FromRequest, HttpRequest, HttpResponse};
 use serde::{Deserialize, Serialize};
 use sled::Db;
 use std::fs;
+use std::ops::Deref;
+
+mod dbu;
 
 #[derive(Serialize, Deserialize)]
 struct FileInfo {
@@ -13,19 +16,20 @@ struct FileInfo {
 #[derive(Serialize, Clone, Deserialize)]
 struct ServerSettings {
     api_keys: Vec<String>,
+    admin_keys: Vec<String>,
     website_name: String,
     https: bool,
 }
 
 #[derive(Deserialize)]
-struct ServeFile {
+struct FilePath {
     folder: String,
     file: String,
 }
 
 #[derive(Deserialize)]
 struct DeleteFile {
-    delete_key: String,
+    del: String,
 }
 
 #[derive(Serialize, Deserialize)]
@@ -49,6 +53,7 @@ fn upload(
     mut parts: awmp::Parts,
     database: actix_web::web::Data<sled::Db>,
     settings: web::Data<ServerSettings>,
+    request: HttpRequest,
 ) -> Result<actix_web::HttpResponse, actix_web::Error> {
     let file_parts = parts
         .files
@@ -57,7 +62,7 @@ fn upload(
         .and_then(|f| f.persist("./uploads").ok())
         .unwrap_or_default();
 
-    let (new_path, uri) = loop {
+    let (new_path, uri, ffn) = loop {
         let file_name = nanoid::generate(6);
         let folder_dir = nanoid::generate(2);
 
@@ -72,7 +77,16 @@ fn upload(
             if !std::path::Path::new(&format!("./uploads/{}", folder_dir)).exists() {
                 fs::create_dir_all(format!("./uploads/{}", folder_dir)).unwrap();
             }
-            break (p, format!("/{}/{}", folder_dir, file_name));
+            break (
+                p,
+                format!("/{}/{}", folder_dir, file_name),
+                format!(
+                    "{}{}.{}",
+                    folder_dir,
+                    file_name,
+                    file_parts.extension().unwrap().to_str().unwrap()
+                ),
+            );
         }
     };
 
@@ -80,17 +94,11 @@ fn upload(
 
     let del_key = nanoid::simple();
 
-    let file_info = FileInfo {
-        delete_key: del_key.clone(),
-        path: new_path,
-    };
+    let ins =
+        dbu::generate_insert_binary(new_path, del_key.clone(), request.connection_info().deref())
+            .unwrap();
 
-    database
-        .insert(
-            del_key.clone().into_bytes(),
-            bincode::serialize(&file_info).unwrap(),
-        )
-        .unwrap();
+    database.insert(ffn.clone().into_bytes(), ins).unwrap();
 
     let resp_json = UploadResp {
         url: format!(
@@ -99,13 +107,19 @@ fn upload(
             uri,
             file_parts.extension().unwrap().to_str().unwrap()
         ),
-        delete_url: format!("https://{}/d/{}", settings.website_name, del_key),
+        delete_url: format!(
+            "https://{}/d{}.{}?del={}",
+            settings.website_name,
+            uri,
+            file_parts.extension().unwrap().to_str().unwrap(),
+            del_key
+        ),
     };
 
     Ok(actix_web::HttpResponse::Ok().json(&resp_json))
 }
 
-fn serve(info: web::Path<ServeFile>) -> actix_web::Result<NamedFile, actix_web::HttpResponse> {
+fn serve(info: web::Path<FilePath>) -> actix_web::Result<NamedFile, actix_web::HttpResponse> {
     let file = format!("./uploads/{}/{}", info.folder, info.file);
 
     match NamedFile::open(file) {
@@ -116,23 +130,31 @@ fn serve(info: web::Path<ServeFile>) -> actix_web::Result<NamedFile, actix_web::
     }
 }
 
-fn delete(delete: web::Path<DeleteFile>, database: web::Data<sled::Db>) -> HttpResponse {
+fn delete(
+    path: web::Path<FilePath>,
+    del: web::Query<DeleteFile>,
+    database: web::Data<sled::Db>,
+) -> HttpResponse {
     let binc = match database
-        .get(delete.delete_key.clone().into_bytes())
+        .get(format!("{}{}", path.folder, path.file))
         .unwrap()
     {
         Some(e) => e,
         None => {
-            return HttpResponse::Unauthorized().body("this is not a valid file delete key");
+            return HttpResponse::Unauthorized().body("this is not a valid file to delete");
         }
     };
+    let data: dbu::FileMetadata = bincode::deserialize(&binc[..]).unwrap();
+
+    if del.del != data.del_key {
+        return HttpResponse::Unauthorized().body("invaild delete key");
+    }
 
     database
-        .remove(delete.delete_key.clone().into_bytes())
+        .remove(format!("{}{}", path.folder, path.file).into_bytes())
         .unwrap();
 
-    let data: FileInfo = bincode::deserialize(&binc[..]).unwrap();
-    let path = std::path::Path::new(&data.path);
+    let path = std::path::Path::new(&data.file_path);
 
     fs::remove_file(path).unwrap();
 
@@ -143,16 +165,39 @@ fn delete(delete: web::Path<DeleteFile>, database: web::Data<sled::Db>) -> HttpR
     HttpResponse::Ok().body("file deleted")
 }
 
-fn stats(databse: web::Data<sled::Db>) -> HttpResponse {
+fn stats(database: web::Data<sled::Db>) -> HttpResponse {
     HttpResponse::Ok().json(Stats {
-        files: databse.len(),
+        files: database.len(),
         version: format!(
-            "{} ({:#?})",
+            "{} {}",
             built_info::PKG_VERSION,
-            built_info::GIT_VERSION
+            built_info::GIT_VERSION.map_or_else(|| "".to_owned(), |v| format!("(git {})", v))
         ),
     })
 }
+
+// TODO: Finish looking glass
+/*fn looking_glass(
+    databse: web::Data<sled::Db>,
+    request: HttpRequest,
+    settings: web::Data<ServerSettings>,
+    info: web::Path<FilePath>,
+) -> HttpResponse {
+    let api_key = match request.headers().get("api") {
+        None => return HttpResponse::Unauthorized().body("no api key supplied"),
+        Some(e) => e,
+    };
+
+    if !settings
+        .admin_keys
+        .iter()
+        .any(|x| x == api_key.to_str().unwrap())
+    {
+        return HttpResponse::Unauthorized().body("bad api key");
+    }
+
+    HttpResponse::Ok().body("coming soon")
+}*/
 
 fn p404() -> &'static str {
     "this resource does not exist."
@@ -176,8 +221,9 @@ fn main() {
             .data(server_settings.clone())
             .data(awmp::Parts::configure(|cfg| cfg.with_temp_dir("./uploads")))
             .route("/u", actix_web::web::post().to(upload))
-            .route("/d/{delete_key}", web::get().to(delete))
+            .route("/d/{folder}/{file}", web::get().to(delete))
             .route("/stats", web::get().to(stats))
+            // .service(web::resource("/lg/{folder}/{file}").route(web::get().to(looking_glass)))
             .service(web::resource("/{folder}/{file}").route(web::get().to(serve)))
             .default_service(web::resource("").route(web::get().to(p404)))
     })
